refactor(frontend): token.validation -> schema
